% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_resources}
\alias{list_resources}
\alias{obrasgov_recursos}
\title{List available ObrasGov API resources}
\usage{
list_resources()

obrasgov_recursos()
}
\value{
A tibble containing each resource name, its corresponding function,
API endpoint, and whether it is paginated.
}
\description{
List available ObrasGov API resources
}
\examples{
list_resources()
}
