% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_get_arquivos_intervencao}
\alias{obrasgov_get_arquivos_intervencao}
\title{Get Intervention Files (Photos and Videos)}
\usage{
obrasgov_get_arquivos_intervencao(idUnico = NULL, page = 0, size = 10)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{page}{Page number (0-based).}

\item{size}{Page size.}
}
\value{
A data frame with intervention files data.
}
\description{
This function retrieves files (photos and videos) related to an intervention from the Obrasgov API.
}
