% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_get_execucao_financeira}
\alias{obrasgov_get_execucao_financeira}
\title{Get Financial Execution Data}
\usage{
obrasgov_get_execucao_financeira(idUnico = NULL, page = 0, size = 10)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{page}{Page number (0-based).}

\item{size}{Page size.}
}
\value{
A data frame with financial execution data.
}
\description{
This function retrieves financial execution data from the Obrasgov API.
}
