% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrasgov_get_geometria.R
\name{obrasgov_get_geometria}
\alias{obrasgov_get_geometria}
\title{Get Georeferencing Data}
\usage{
obrasgov_get_geometria(
  idUnico = NULL,
  showProgress = TRUE,
  max_pages = Inf,
  parallel = FALSE,
  workers = 2
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{max_pages}{An integer to limit the maximum number of pages to download. Defaults to \code{Inf} (all pages).}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}

\item{tamanhoDaPagina}{NSet the page size. A larger value reduces the number of API calls.}
}
\value{
A data frame with georeferencing data.
}
\description{
This function retrieves georeferencing data from the Obrasgov API.
}
