% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_get_projetos_investimento}
\alias{obrasgov_get_projetos_investimento}
\title{Get Investment Projects}
\usage{
obrasgov_get_projetos_investimento(
  idUnico = NULL,
  situacao = NULL,
  codigoOrganizacao = NULL,
  nomeOrganizacao = NULL,
  uf = NULL,
  municipio = NULL,
  eixo = NULL,
  subEixo = NULL,
  tipoIntervencao = NULL,
  naturezaIntervencao = NULL,
  fonteRecurso = NULL,
  dataInicio = NULL,
  dataFim = NULL,
  page = 0,
  size = 10
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{situacao}{Situation of the intervention (e.g., "Em Execução", "Paralisada", "Cancelada", "Castrada", "Concluída" or "Inacabada").}

\item{codigoOrganizacao}{Organization code participating in the intervention.}

\item{nomeOrganizacao}{Organization name participating in the intervention.}

\item{uf}{Main UF of the intervention (e.g., "DF").}

\item{municipio}{Main municipality of the intervention (e.g., "Brasília").}

\item{eixo}{Investment project axis (e.g., "EDUCAÇÃO").}

\item{subEixo}{Investment project sub-axis (e.g., "EDUCAÇÃO BÁSICA").}

\item{tipoIntervencao}{Type of intervention (e.g., "OBRAS").}

\item{naturezaIntervencao}{Nature of intervention (e.g., "CONSTRUÇÃO").}

\item{fonteRecurso}{Source of resource (e.g., "ORÇAMENTO GERAL DA UNIÃO").}

\item{dataInicio}{Start date of the intervention (format: yyyy-MM-dd).}

\item{dataFim}{End date of the intervention (format: yyyy-MM-dd).}

\item{page}{Page number (0-based).}

\item{size}{Page size.}
}
\value{
A data frame with investment projects.
}
\description{
This function retrieves investment projects from the Obrasgov API.
}
