% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_get_saldo_contabil}
\alias{obrasgov_get_saldo_contabil}
\title{Get Financial Balance Data}
\usage{
obrasgov_get_saldo_contabil(idUnico = NULL, page = 0, size = 10)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{page}{Page number (0-based).}

\item{size}{Page size.}
}
\value{
A data frame with financial balance data.
}
\description{
This function retrieves financial balance data from the Obrasgov API.
}
