% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geom.R
\name{read_geom}
\alias{read_geom}
\title{Get Georeferencing Data}
\usage{
read_geom(idUnico = NULL, showProgress = TRUE)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}
}
\value{
A data frame with georeferencing data.
}
\description{
This function retrieves Projects' georeferencing data from the Obrasgov.BR API.
}
