% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_media.R
\name{read_media}
\alias{read_media}
\title{Get Media Files (Photos and Videos)}
\usage{
read_media(idUnico = NULL, showProgress = TRUE)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}
}
\value{
A data frame with intervention files data.
}
\description{
This function retrieves files (photos and videos) related to an project from the Obrasgov API.
}
