% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physical_progress.R
\name{read_physical_progress}
\alias{read_physical_progress}
\title{Get Physical Execution Data}
\usage{
read_physical_progress(
  idUnico = NULL,
  situacao = NULL,
  showProgress = TRUE,
  max_pages = Inf,
  parallel = FALSE,
  workers = 2
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{situacao}{Situation of the intervention ("Em Execução", "Paralisada", "Cancelada", "Cadastrada", "Concluída" or "Inacabada").}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{max_pages}{An integer to limit the maximum number of pages to download. Defaults to \code{Inf} (all pages).}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}

\item{tamanhoDaPagina}{Set the page size. A larger value reduces the number of API calls. HIGH NUMBERS (>30) RESULT IN PAGING LOOP PROBLEMS!}
}
\value{
A data frame with physical execution data.
}
\description{
This function retrieves physical execution data from the Obrasgov API.
}
