% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{read_project}
\alias{read_project}
\title{Get Investment Projects}
\usage{
read_project(
  idUnico = NULL,
  situacao = NULL,
  codigoOrganizacao = NULL,
  nomeOrganizacao = NULL,
  uf = NULL,
  dataCadastro = NULL,
  natureza = NULL,
  tamanhoDaPagina = 10,
  geometry = TRUE,
  showProgress = TRUE,
  max_pages = Inf,
  parallel = FALSE,
  workers = 2
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{situacao}{Situation of the intervention (e.g., "Em Execução").}

\item{codigoOrganizacao}{Organization code participating in the intervention.}

\item{nomeOrganizacao}{Organization name participating in the intervention.}

\item{uf}{Main UF of the intervention (e.g., "DF").}

\item{dataCadastro}{Intervention Registration Date "YYYY-MM-DD".}

\item{natureza}{Intervation nature (e.g, "Obra").}

\item{tamanhoDaPagina}{Set the page size. A larger value reduces the number of API calls.}

\item{geometry}{Converts WKB code to Geometry.  Defaults to \code{TRUE}. Turn FALSE if you feel it is breaking the downloading.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{max_pages}{Maximum number of pages to download. Change to \code{Inf} to download all pages.}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}
}
\value{
A data frame with investment projects.
}
\description{
This function retrieves investment projects from the Obrasgov API.
It supports filtering, pagination control, and parallel downloads.
}
