% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_api_request}
\alias{obrasgov_api_request}
\title{Base function to make API requests to Obrasgov}
\usage{
obrasgov_api_request(path, query_params = list())
}
\arguments{
\item{path}{The API endpoint path (e.g., "/projeto-investimento").}

\item{query_params}{A list of query parameters.}
}
\value{
A data frame with the API response.
}
\description{
Base function to make API requests to Obrasgov
}
