% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrasgov_get_execucao_fisica.R
\name{obrasgov_get_arquivos_intervencao}
\alias{obrasgov_get_arquivos_intervencao}
\title{Get Intervention Files (Photos and Videos)}
\usage{
obrasgov_get_arquivos_intervencao(idUnico = NULL, showProgress = TRUE)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}
}
\value{
A data frame with intervention files data.
}
\description{
This function retrieves files (photos and videos) related to an intervention from the Obrasgov API.
}
