% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrasgov_get_execucao_fisica.R
\name{obrasgov_get_execucao_fisica}
\alias{obrasgov_get_execucao_fisica}
\title{Get Physical Execution Data}
\usage{
obrasgov_get_execucao_fisica(
  idUnico = NULL,
  situacao = NULL,
  showProgress = TRUE
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{situacao}{Situation of the intervention (e.g., "Em Execução").}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{tamanhoDaPagina}{NSet the page size. A larger value reduces the number of API calls.}

\item{max_pages}{An integer to limit the maximum number of pages to download. Defaults to \code{Inf} (all pages).}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}
}
\value{
A data frame with physical execution data.
}
\description{
This function retrieves physical execution data from the Obrasgov API.
}
