% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obrasgov_get_geometria}
\alias{obrasgov_get_geometria}
\title{Get Georeferencing Data}
\usage{
obrasgov_get_geometria(idUnico = NULL, page = 0, size = 10)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{page}{Page number (0-based).}

\item{size}{Page size.}
}
\value{
A data frame with georeferencing data.
}
\description{
This function retrieves georeferencing data from the Obrasgov API.
}
