% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrasgov_get_projetos_investimento.R
\name{obrasgov_get_projetos_investimento}
\alias{obrasgov_get_projetos_investimento}
\title{Get Investment Projects}
\usage{
obrasgov_get_projetos_investimento(
  idUnico = NULL,
  situacao = NULL,
  codigoOrganizacao = NULL,
  nomeOrganizacao = NULL,
  uf = NULL,
  dataCadastro = NULL,
  natureza = NULL,
  tamanhoDaPagina = 20,
  showProgress = TRUE,
  max_pages = Inf,
  parallel = FALSE,
  workers = 2
)
}
\arguments{
\item{idUnico}{Unique identifier of the intervention.}

\item{situacao}{Situation of the intervention (e.g., "Em Execução").}

\item{codigoOrganizacao}{Organization code participating in the intervention.}

\item{nomeOrganizacao}{Organization name participating in the intervention.}

\item{uf}{Main UF of the intervention (e.g., "DF").}

\item{dataCadastro}{Intervention Registration Date "YYYY-MM-DD".}

\item{natureza}{Intervation nature (e.g, "Obra").}

\item{tamanhoDaPagina}{NSet the page size. A larger value reduces the number of API calls.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{max_pages}{An integer to limit the maximum number of pages to download. Defaults to \code{Inf} (all pages).}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}
}
\value{
A data frame with investment projects.
}
\description{
This function retrieves investment projects from the Obrasgov API.
It supports filtering, pagination control, and parallel downloads.
}
