% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrasgov_get_execucao_financeira.R
\name{obrasgov_get_saldo_contabil}
\alias{obrasgov_get_saldo_contabil}
\title{Get Financial Balance Data}
\usage{
obrasgov_get_saldo_contabil(
  ugEmitente = NULL,
  nrNotaEmpenho = NULL,
  tamanhoDaPagina = 20,
  showProgress = TRUE,
  max_pages = Inf,
  parallel = FALSE,
  workers = 2
)
}
\arguments{
\item{ugEmitente}{UG Emitente}

\item{nrNotaEmpenho}{Número do Empenho.}

\item{tamanhoDaPagina}{NSet the page size. A larger value reduces the number of API calls.}

\item{showProgress}{A logical value to control whether the progress bar is displayed.}

\item{max_pages}{An integer to limit the maximum number of pages to download. Defaults to \code{Inf} (all pages).}

\item{parallel}{A logical value to enable parallel processing for downloads. Defaults to \code{FALSE}.}

\item{workers}{An integer specifying the number of parallel workers (cores) to use when \code{parallel = TRUE}.}
}
\value{
A data frame with financial balance data.
}
\description{
This function retrieves financial balance data from the Obrasgov API.
}
